fix overflow n_reads in show_progress_until_done

+++ src/aln.cpp
@@ -713,7 +713,7 @@ static inline void rescue_mate(
     alignment &sam_aln,
     float mu,
     float sigma,
-    unsigned int &tot_rescued,
+    uint64_t &tot_rescued,
     int k
 ) {
     int a, b, ref_start,ref_len,ref_end;
